feat: Add solution for 206 reverse linked list
This commit adds a solution for the "Reverse Linked List" problem on LeetCode. The solution uses the iterative approach to reverse a singly-linked list. It initializes two pointers, `prev` and `curr`, and iterates through the list, updating the `next` pointers to reverse the list. Finally, it returns the new head of the reversed list.